fix macos: the buffer returned by getattrlist was interpreted in the wrong way. The count at the beginning of the buffer is 32 bit even on 64 bit systems. This resulted in always returning true, i.e. the file system was considered case-insensitive.

+++ src/clib/lm_fs_case_sensitive.c
@@ -50,7 +50,7 @@
 #include <errno.h>
 
 typedef struct vol_caps_buf {
-    unsigned long size;
+    u_int32_t size;
     vol_capabilities_attr_t caps;
 } vol_caps_buf_t;
